runtime: (refactor) Move StreamRuntime into impl.stream package
