+++ src/main/java/com/abusalimov/mrcalc/runtime/impl/stream/DoubleArraySequence.java
@@ -1,4 +1,7 @@
-package com.abusalimov.mrcalc.runtime;
+package com.abusalimov.mrcalc.runtime.impl.stream;
+
+import com.abusalimov.mrcalc.runtime.DoubleSequence;
+import com.abusalimov.mrcalc.runtime.Sequence;
 
 import java.util.AbstractList;
 import java.util.Arrays;
