+++ src/main/java/com/abusalimov/mrcalc/runtime/impl/stream/LongArrayRange.java
@@ -1,4 +1,4 @@
-package com.abusalimov.mrcalc.runtime;
+package com.abusalimov.mrcalc.runtime.impl.stream;
 
 /**
  * A simple and naive implementation of range of longs.
